requirements.txt
Flask
requests
mysql-connector-python
gunicorn
flask-login
flask-sqlalchemy
flask-bcrypt
python-dotenv
